Checkpoint, shadow sample
