+++ samples/CMakeLists.txt
@@ -24,5 +24,6 @@ add_subdirectory(pub_sub/basic_pub_sub)
 add_subdirectory(pub_sub/cycle_pub_sub)
 add_subdirectory(secure_tunneling/secure_tunnel)
 add_subdirectory(secure_tunneling/tunnel_notification)
+add_subdirectory(shadow/deprecated/shadow_sync)
+add_subdirectory(shadow/deprecated/mqtt5_shadow_sync)
 add_subdirectory(shadow/shadow_sync)
-add_subdirectory(shadow/mqtt5_shadow_sync)
