+++ samples/README.md
@@ -10,7 +10,7 @@
     + [HTTP Proxy](./mqtt5/mqtt5_pubsub/README.md#http-proxy)
 * [Mqtt5 Shared Subscription](./mqtt5/mqtt5_shared_subscription/README.md)
 * [Mqtt5 Jobs](./jobs/mqtt5_job_execution/README.md)
-* [Mqtt5 Shadow](./shadow/mqtt5_shadow_sync/README.md)
+* [Shadow](./shadow/shadow_sync/README.md)
 * [Mqtt5 Fleet Provisioning](./fleet_provisioning/mqtt5_fleet_provisioning/README.md)
 ## MQTT311 Samples
 * [Basic Pub-Sub](./pub_sub/basic_pub_sub/README.md)
@@ -22,7 +22,6 @@
 * [Windows Certificate MQTT Connect](./mqtt/windows_cert_connect/README.md)
 * [Custom Authorizer Connect](./mqtt/custom_authorizer_connect/README.md)
 * [Cognito Connect](./mqtt/cognito_connect/README.md)
-* [Shadow](./shadow/shadow_sync/README.md)
 * [Jobs](./jobs/job_execution/README.md)
 * [Fleet provisioning](./fleet_provisioning/fleet_provisioning/README.md)
 ## Other Samples
