+++ samples/shadow/deprecated/mqtt5_shadow_sync/CMakeLists.txt
@@ -4,8 +4,8 @@ project(mqtt5-shadow-sync CXX)
 
 file(GLOB SRC_FILES
        "*.cpp"
-       "../../utils/CommandLineUtils.cpp"
-       "../../utils/CommandLineUtils.h"
+       "../../../utils/CommandLineUtils.cpp"
+       "../../../utils/CommandLineUtils.h"
 )
 
 add_executable(${PROJECT_NAME} ${SRC_FILES})
