+++ samples/shadow/deprecated/mqtt5_shadow_sync/README.md
