+++ samples/shadow/deprecated/shadow_sync/CMakeLists.txt
@@ -1,11 +1,11 @@
-cmake_minimum_required(VERSION 3.1)
+cmake_minimum_required(VERSION 3.9)
 # note: cxx-17 requires cmake 3.8, cxx-20 requires cmake 3.12
-project(shadowv2 CXX)
+project(shadow-sync CXX)
 
 file(GLOB SRC_FILES
        "*.cpp"
-       "../../utils/CommandLineUtils.cpp"
-       "../../utils/CommandLineUtils.h"
+       "../../../utils/CommandLineUtils.cpp"
+       "../../../utils/CommandLineUtils.h"
 )
 
 add_executable(${PROJECT_NAME} ${SRC_FILES})
