+++ samples/shadow/shadow_sync/CMakeLists.txt
@@ -1,6 +1,6 @@
-cmake_minimum_required(VERSION 3.9)
+cmake_minimum_required(VERSION 3.1)
 # note: cxx-17 requires cmake 3.8, cxx-20 requires cmake 3.12
-project(shadow-sync CXX)
+project(shadowv2 CXX)
 
 file(GLOB SRC_FILES
        "*.cpp"
